@@ -4,20 +4,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-import eu.su.mas.dedaleEtu.mas.agents.dummies.DummyCollectorAgent;
-import eu.su.mas.dedaleEtu.mas.agents.dummies.DummyMovingAgent;
-import eu.su.mas.dedaleEtu.mas.agents.dummies.DummyTankerAgent;
-import eu.su.mas.dedaleEtu.mas.agents.dummies.explo.ExploreCoopAgent;
-import eu.su.mas.dedaleEtu.mas.agents.dummies.explo.ExploreSoloAgent;
 import eu.su.mas.dedale.mas.AbstractDedaleAgent;
 import eu.su.mas.dedale.mas.agents.GateKeeperAgent;
-import eu.su.mas.dedale.mas.agents.dedaleDummyAgents.DummyWumpusShift;
-import eu.su.mas.dedale.mas.agents.dedaleDummyAgents.DummyWumpusShift2;
-import eu.su.mas.dedale.mas.agents.dedaleDummyAgents.Controlled.ControlledAgent;
 import eu.su.mas.dedaleEtu.mas.agents.treasureHunt.CollectorAgent;
 import eu.su.mas.dedaleEtu.mas.agents.treasureHunt.ExplorerAgent;
 import eu.su.mas.dedaleEtu.mas.agents.treasureHunt.MessengerAgent;
-import eu.su.mas.dedaleEtu.mas.agents.treasureHunt.TreasureExplorer;
 import jade.core.Profile;
 import jade.core.ProfileImpl;
 import jade.core.Runtime;
@@ -295,44 +286,44 @@ public class Principal {
 		agentName = "Explo4";
 		ag = createNewDedaleAgent(c, agentName, ExplorerAgent.class.getName(), entityParameters);
 		agentList.add(ag);
-		//Initialize collectors
-		agentName = "Collect1";
-		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Collect2";
-		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Collect3";
-		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Collect4";
-		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		//Initialize collectors
-		agentName = "Msg1";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg2";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg3";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg4";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg5";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg6";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg7";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
-		agentName = "Msg8";
-		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
-		agentList.add(ag);
+//		//Initialize collectors
+//		agentName = "Collect1";
+//		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Collect2";
+//		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Collect3";
+//		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Collect4";
+//		ag = createNewDedaleAgent(c, agentName, CollectorAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		//Initialize collectors
+//		agentName = "Msg1";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg2";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg3";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg4";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg5";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg6";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg7";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
+//		agentName = "Msg8";
+//		ag = createNewDedaleAgent(c, agentName, MessengerAgent.class.getName(), entityParameters);
+//		agentList.add(ag);
 
 		System.out.println("Agents created...");
 		return agentList;
